+++ src/migrations/M240124120902Nullable.php
@@ -39,8 +39,4 @@ class M240124120902Nullable extends Migration
             }
         });
     }
-
-    public function safeDown(): void
-    {
-    }
 }